login procedure to open browser on login failure when headless
Fixed a bug in the login procedure, so that when login fails due to Captcha, the browser will be displayed to the user for them to complete the Captcha. Also automatically close the browser after logging in to reduce strain on the user's computer. The browser is not needed after login. There is still an option -- CloseDriverAfterComplete -- that can be set to keep the browser open after successful login.